Fix des formulaires de modification de longueur et de voie - Fix de l'affichage des spots et du compteur de spots dans la fiche d'un topo - Ajout de TopoDto pour le traitement des données du formulaire de création de topo (en cours)

+++ src/main/java/fr/axelallain/dto/TopoDto.java
@@ -12,7 +12,7 @@ public class TopoDto {
 	
 	public String lieu;
 	
-	List<Integer> spotsId;
+	List<Long> spotsId;
 
 	public Long getUtilisateurId() {
 		return utilisateurId;
@@ -46,11 +46,11 @@ public class TopoDto {
 		this.lieu = lieu;
 	}
 
-	public List<Integer> getSpotsId() {
+	public List<Long> getSpotsId() {
 		return spotsId;
 	}
 
-	public void setSpotsId(List<Integer> spotsId) {
+	public void setSpotsId(List<Long> spotsId) {
 		this.spotsId = spotsId;
 	}
